refactor: group shared content components
