+++ src/shared/components/content/Card.tsx
