+++ src/shared/components/content/Logo.tsx
