+++ src/shared/components/content/Score.tsx
