+++ src/shared/components/content/Section.tsx
