+++ src/shared/components/content/Style.tsx
