+++ src/shared/components/content/TagList.stories.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import type { Meta, StoryObj } from "@storybook/react";
 
 import { Tag } from "@src/shared/components/forms/Tag";
-import { TagList as Template } from "@src/shared/components/TagList";
+import { TagList as Template } from "@src/shared/components/content/TagList";
 import * as fixture from "@src/shared/storybook/fixture";
 
 const meta = {
