+++ src/shared/components/index.ts
@@ -1,7 +1,7 @@
 export * from "@src/shared/components/forms/Button";
-export * from "@src/shared/components/Card";
-export * from "@src/shared/components/Code";
-export * from "@src/shared/components/Description";
+export * from "@src/shared/components/content/Card";
+export * from "@src/shared/components/content/Code";
+export * from "@src/shared/components/content/Description";
 export * from "@src/shared/components/Feedback";
 export * from "@src/shared/components/forms/Form";
 export * from "@src/shared/components/forms/FormItem";
@@ -10,19 +10,19 @@ export * from "@src/shared/components/layout/Header";
 export * from "@src/shared/components/forms/Input";
 export * from "@src/shared/components/layout/Layout";
 export * from "@src/shared/components/layout/List";
-export * from "@src/shared/components/Logo";
+export * from "@src/shared/components/content/Logo";
 export * from "@src/shared/components/layout/Main";
-export * from "@src/shared/components/Math";
+export * from "@src/shared/components/content/Math";
 export * from "@src/shared/components/layout/Outer";
 export * from "@src/shared/components/Overlay";
-export * from "@src/shared/components/Score";
-export * from "@src/shared/components/Section";
+export * from "@src/shared/components/content/Score";
+export * from "@src/shared/components/content/Section";
 export * from "@src/shared/components/forms/Select";
 export * from "@src/shared/components/forms/Slider";
-export * from "@src/shared/components/Style";
+export * from "@src/shared/components/content/Style";
 export * from "@src/shared/components/forms/Switch";
 export * from "@src/shared/components/forms/Tag";
-export * from "@src/shared/components/TagList";
+export * from "@src/shared/components/content/TagList";
 export * from "@src/shared/components/forms/Textarea";
-export * from "@src/shared/components/Title";
+export * from "@src/shared/components/content/Title";
 export * from "@src/shared/components/forms/Upload";
