+++ src/shared/components/layout/Header.tsx
@@ -2,7 +2,7 @@ import cx from "classnames";
 import * as React from "react";
 import { IconContext } from "react-icons";
 import { AiFillSetting, AiOutlineUpload, AiOutlineSun, AiFillMoon } from "react-icons/ai";
-import { Logo } from "@src/shared/components/Logo";
+import { Logo } from "@src/shared/components/content/Logo";
 
 export interface HeaderProps {
   fixed?: boolean;
