+++ src/shared/components/layout/List.stories.tsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import type { Meta, StoryObj } from "@storybook/react";
 
-import { Card } from "@src/shared/components/Card";
+import { Card } from "@src/shared/components/content/Card";
 import { List as Template } from "@src/shared/components/layout/List";
 
 const meta = {
